6022BE: Make samplerate adjust to chosen timebase.
The product of samplerate and timebase (x10) specifies the number of samples
needed to fill up the screen horizontally.
Based on the record length, the system was able to configure a timebase
when the samplerate was changed.

Code exist indicating this should also work in the opposite direction. I.e.
when changing timebase, the samplerate should adjust.

This has been fixed to also work now for 6022BE